Reorganised code
